Transformations: Separate `transformations.inline` into sub-package
